rename common to fixtures and refactor all usages
