+++ tests/test_metric.py
@@ -9,7 +9,7 @@ from pie_core import (
     DocumentMetric,
     annotation_field,
 )
-from tests.common.types import LabeledSpan, TextBasedDocument
+from tests.fixtures.types import LabeledSpan, TextBasedDocument
 
 
 @pytest.fixture
